+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - bugfix: clear test results on new run
 - added comment support
 - added mocha support
+- added success/fail badge support at top
 
 0.1.0 / 2015-07-25
 ------------------